+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [unreleased]
 ### changed
 * bumped msrv to 1.66.1
+* will now pre-allocate memory for config file based on its metadata
 
 ## [0.7.0] - 2023-01-25
 ### added